+++ components/Navbar/index.ts
@@ -1,0 +1,3 @@
+
+export { Navbar } from "./Navbar.js"
+export { NavbarLogo } from "./NavbarLogo.js"